Address clang tidy issues, optimize line connectivity and fix line intersection math

+++ include/tag_detection/internal/tag25h7.h
@@ -6,7 +6,7 @@ namespace tag_detection {
 
 // Tag bits are read like a book; left to right, top to bottom. Most significant bit is top left.
 
-const std::vector<unsigned long long> t25h7{
+const std::vector<uint64_t> t25h7{
     0x4b770dLL,  0x11693e6LL, 0x1a599abLL, 0xc3a535LL,  0x152aafaLL, 0xaccd98LL,  0x1cad922LL,
     0x2c2fadLL,  0xbb3572LL,  0x14a3b37LL, 0x186524bLL, 0xc99d4cLL,  0x23bfeaLL,  0x141cb74LL,
     0x1d0d139LL, 0x1670aebLL, 0x851675LL,  0x150334eLL, 0x6e3ed8LL,  0xfd449dLL,  0xaa55ecLL,
